Refactor and add docs to pyro.contrib.cevae
