+++ docs/source/conf.py
@@ -42,6 +42,7 @@ extensions = [
     'sphinx.ext.ifconfig',  #
     'sphinx.ext.viewcode',  #
     'sphinx.ext.githubpages',  #
+    'sphinx.ext.graphviz',  #
     'sphinx.ext.autodoc',
     'sphinx.ext.doctest',
 ]
