+++ docs/source/contrib.cevae.rst
@@ -1,0 +1,52 @@
+Causal Effect VAE
+=================
+
+.. automodule:: pyro.contrib.cevae
+
+CEVAE Class
+-----------
+.. autoclass:: pyro.contrib.cevae.CEVAE
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
+
+CEVAE Components
+----------------
+.. autoclass:: pyro.contrib.cevae.Model
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
+
+.. autoclass:: pyro.contrib.cevae.Guide
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
+
+.. autoclass:: pyro.contrib.cevae.TraceCausalEffect_ELBO
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
+
+Utilities
+---------
+.. autoclass:: pyro.contrib.cevae.FullyConnected
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
+
+.. autoclass:: pyro.contrib.cevae.DiagNormalNet
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
+
+.. autoclass:: pyro.contrib.cevae.BernoulliNet
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
